backend/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
flask-restx==1.0.3
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.1.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.1
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
pkgutil-resolve-name==1.3.10
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.10.0
